Move to header file: time and text output
