+++ Qubic.vcxproj
@@ -32,11 +32,13 @@
   <ItemGroup>
     <ClInclude Include="four_q.h" />
     <ClInclude Include="kangaroo_twelve.h" />
+    <ClInclude Include="text_output.h" />
     <ClInclude Include="private_settings.h" />
     <ClInclude Include="public_settings.h" />
     <ClInclude Include="qpi.h" />
     <ClInclude Include="qubics\Qx.h" />
     <ClInclude Include="smart_contracts.h" />
+    <ClInclude Include="time.h" />
     <ClInclude Include="uefi.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
