+++ Qubic.vcxproj.filters
@@ -12,5 +12,7 @@
     <ClInclude Include="uefi.h" />
     <ClInclude Include="kangaroo_twelve.h" />
     <ClInclude Include="four_q.h" />
+    <ClInclude Include="text_output.h" />
+    <ClInclude Include="time.h" />
   </ItemGroup>
 </Project>
